moved expectation function, changed test to work with the move, and updated pyscf version
