+++ .github/workflows/tests.yml
@@ -14,5 +14,5 @@ jobs:
     with:
       os: ${{ matrix.os }}
       python-version: ${{ matrix.python-version }}
-      poetry-extras: "--with docs,tests,analysis --all-extras"
+      poetry-extras: --with docs,tests,analysis --all-extras
     secrets: inherit
