+++ pyproject.toml
@@ -18,7 +18,7 @@ packages = [{ include = "qibochem", from = "src" }]
 python = "^3.9"
 qibo = "^0.1.10"
 openfermion = "^1.5"
-pyscf = "^2.2.1"
+pyscf = "^2.3.0"
 
 [build-system]
 requires = ["poetry-core"]
